requirements.txt
opencv-python
Flask
pillow
ultralytics